clean up TstUtils
- removed unused #copyItemWithMeta (should be replaced by GTUtility#copyMetaData)
- removed unused #copyAmountUnlimited
- replaced #copyAmountFluid with GTUtility#copyAmount(int, FluidStack)
- added javadocs for #setStackSize(ItemStack, int) and #setStackSize(FluidStack, int)
- added javadocs for #removeIntegratedCircuitFromStacks(ItemStack[])
- moved #mergeSameFluid and #mergeSameItem to MiracleTopRecipePool where there is the only place can use them with logical flaws lol
